pyproject.toml
[project]
name = "ciri"
description = "CIRI Copilot — a local AI agent CLI with skills, toolkits, and subagents"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE.md" }
authors = [
    { name = "Aditya Mishra", email = "adimis.ai.001@gmail.com" }
]
keywords = [
    "ai", "agent", "cli", "langgraph", "llm", "copilot",
    "automation", "skills", "toolkits", "subagents", "langchain"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]

dependencies = [
    "typer[all]",
    "rich",
    "anyio",
    "httpx",
    "tenacity",
    "orjson",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "openai",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "langgraph",
    "deepagents",
    "crawl4ai",
    "duckduckgo-search",
    "ddgs",
    "sqlite-vec",
    "sqlalchemy",
    "langchain-unstructured",
    "langchain-mcp-adapters",
    "langgraph-checkpoint-sqlite",
    "structlog",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "packaging",
    "platformdirs",
    "importlib-metadata; python_version<'3.13'",
    "black>=26.1.0",
    "langgraph-swarm>=0.1.0",
    "lxml>=5.4.0",
    "playwright>=1.58.0",
    "prompt-toolkit>=3.0.52",
    "asgiref>=3.11.1",
    "langchain-litellm>=0.5.1",
    "langchain-anthropic>=1.3.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-mistralai>=1.1.1",
    "langchain-cohere>=0.5.0",
    "langchain-together>=0.0.2.post1",
    "langchain-fireworks>=1.1.0",
    "langchain-ollama>=1.0.1",
    "langchain-deepseek>=1.0.1",
    "langchain-openrouter>=0.0.2",
]

[project.optional-dependencies]
encrypted = [
    "pysqlcipher3",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "black",
    "mypy",
    "types-PyYAML",
    "types-requests",
    "build",
    "twine",
    "pyinstaller",
    "cyclonedx-bom",
    "pip-audit",
]

[project.urls]
Homepage = "https://github.com/adimis-ai/ciri"
Repository = "https://github.com/adimis-ai/ciri"
Issues = "https://github.com/adimis-ai/ciri/issues"
Changelog = "https://github.com/adimis-ai/ciri/releases"

[project.scripts]
ciri = "ciri.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "ciri/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["ciri"]
include = [
    "ciri/**/*.py",
    "ciri/**/*.md",
    "ciri/**/*.yaml",
    "ciri/**/*.yml",
    "ciri/**/*.json",
    "ciri/**/*.txt",
    "ciri/**/*.xsd",
    "ciri/**/*.ttf",
    "ciri/**/*.xml",
    "ciri/**/*.sh",
    "ciri/**/*.pdf",
    "ciri/**/*.js",
    "ciri/**/*.html",
    "ciri/**/*.gz",
]

[tool.hatch.build.targets.sdist]
include = [
    "ciri/",
    "tests/",
    "README.md",
    "LICENSE.md",
    "CONTRIBUTING.md",
    "pyproject.toml",
]
